Added Titlepage for IMT Graz University of Techn.
Titlepage is based on the guidelines published on the institutes
website. Adress of the institute is hardcoded since this title-
page only applies to this specific institute anyway. Also added
\myinstitutehead and \myevaluator to the userdata.tex which might
be useful for other titlepages as well. The guidlines are kind of
weird, for instance: Why would they put the date into brackets?